Move connection settings into standalone component.
